Make DefaultImageLoader testable
The caches, http client and dispatcher are now injected, so the loader no
longer reads the application context while initialising. Requiring it
lazily also means a host app that disables androidx.startup no longer
poisons the class.
